+++ nolds/measures.py
@@ -681,7 +681,7 @@ def plot_dists(dists, tolerance, m, title=None, fname=None):
 
 
 def sampen(data, emb_dim=2, tolerance=None, dist=rowwise_chebyshev,
-           debug_plot=False, debug_data=False, plot_file=None):
+           closed=False, debug_plot=False, debug_data=False, plot_file=None):
   """
   Computes the sample entropy of the given data.
 
@@ -726,6 +726,10 @@ def sampen(data, emb_dim=2, tolerance=None, dist=rowwise_chebyshev,
       distance function used to calculate the distance between template
       vectors. Sampen is defined using ``rowwise_chebyshev``. You should only
       use something else, if you are sure that you need it.
+    closed (boolean):
+      if True, will check for vector pairs whose distance is in the closed
+      interval [0, r] (less or equal to r), otherwise the open interval
+      [0, r) (less than r) will be used
     debug_plot (boolean):
       if True, a histogram of the individual distances for m and m+1
     debug_data (boolean):
@@ -739,6 +743,9 @@ def sampen(data, emb_dim=2, tolerance=None, dist=rowwise_chebyshev,
     float:
       the sample entropy of the data (negative logarithm of ratio between
       similar template vectors of length emb_dim + 1 and emb_dim)
+    [c_m, c_m1]:
+      list of two floats: count of similar template vectors of length emb_dim
+      (c_m) and of length emb_dim + 1 (c_m1)
     [float list, float list]:
       Lists of lists of the form ``[dists_m, dists_m1]`` containing the
       distances between template vectors for m (dists_m)
@@ -776,20 +783,38 @@ def sampen(data, emb_dim=2, tolerance=None, dist=rowwise_chebyshev,
     # successively calculate distances between each pair of template vectors
     for i in range(len(tVecsM) - 1):
       dsts = dist(tVecsM[i + 1:], tVecsM[i])
-      if debug_plot:
+      if debug_plot or debug_data:
         plot_data[-1].extend(dsts)
       # count how many distances are smaller than the tolerance
-      counts[-1] += np.sum(dsts < tolerance)
-  if counts[1] == 0:
-    # log would be infinite => cannot determine saen
-    saen = np.inf
-  else:
+      if closed:
+        counts[-1] += np.sum(dsts <= tolerance)
+      else:
+        counts[-1] += np.sum(dsts < tolerance)
+  if counts[0] > 0 and counts[1] > 0:
     saen = -np.log(1.0 * counts[1] / counts[0])
+  else:
+    # log would be infinite or undefined => cannot determine saen
+    zcounts = []
+    if counts[0] == 0:
+      zcounts.append("emb_dim")
+    if counts[1] == 0:
+      zcounts.append("emb_dim + 1")
+    warnings.warn(
+      "Zero vectors are within tolerance for %s. " \
+      + "Consider raising the tolerance parameter to avoid %s result." \
+      % (" and ".join(zcounts), "NaN" if len(zcounts) == 2 else "inf")
+    )
+    if counts[0] == 0 and counts[1] == 0:
+      saen = np.nan
+    elif counts[0] == 0:
+      saen = -np.inf
+    else:
+      saen = np.inf
   if debug_plot:
     plot_dists(plot_data, tolerance, m, title="sampEn = {:.3f}".format(saen),
                fname=plot_file)
   if debug_data:
-    return (saen, plot_data)
+    return (saen, counts, plot_data)
   else:
     return saen