+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- new parameter `closed` for `sampen` to determine if `<` or `<=` should be used for checking if the distance between a vector pair is within the tolerance
 - new measures
   - `mfhurst_b` calculates the multifractal "generalized" Hurst exponent according to Barabasi et al.
   - `mfhurst_dm` calculates the multifractal "generalized" Hurst exponent according to Di Matteo et al.
@@ -20,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- `debug_data` for `sampen` now also contains counts of similar vectors
+- `sampen` now issues a warning if one or both count variables are zero
 - uses `np.float64` as standard `dtype` instead of `"float32"`
 
 ### Fixed